+++ espnet/nets/pytorch_backend/e2e_tts_fastspeech.py
@@ -13,8 +13,11 @@ from espnet.asr.asr_utils import torch_load
 from espnet.nets.pytorch_backend.e2e_tts_tacotron2 import make_non_pad_mask
 from espnet.nets.pytorch_backend.e2e_tts_transformer import Transformer
 from espnet.nets.pytorch_backend.e2e_tts_transformer import TTSPlot
+from espnet.nets.pytorch_backend.fastspeech.duration_calculator import DurationCalculator
+from espnet.nets.pytorch_backend.fastspeech.duration_predictor import DurationPredictor
+from espnet.nets.pytorch_backend.fastspeech.duration_predictor import DurationPredictorLoss
+from espnet.nets.pytorch_backend.fastspeech.length_regularizer import LengthRegularizer
 from espnet.nets.pytorch_backend.nets_utils import make_pad_mask
-from espnet.nets.pytorch_backend.nets_utils import pad_list
 from espnet.nets.pytorch_backend.transformer.attention import MultiHeadedAttention
 from espnet.nets.pytorch_backend.transformer.embedding import PositionalEncoding
 from espnet.nets.pytorch_backend.transformer.embedding import ScaledPositionalEncoding
@@ -497,187 +500,3 @@ class FeedForwardTransformer(TTSInterface, torch.nn.Module):
             plot_keys += ["encoder_alpha", "decoder_alpha"]
 
         return plot_keys
-
-
-class DurationPredictorLoss(torch.nn.Module):
-    """Duration predictor loss module
-
-    :param float offset: offset value to avoid nan in log domain
-    """
-
-    def __init__(self, offset=1.0):
-        super(DurationPredictorLoss, self).__init__()
-        self.criterion = torch.nn.MSELoss()
-        self.offset = offset
-
-    def forward(self, outputs, targets):
-        """Calculate loss value
-
-        :param torch.Tensor outputs: prediction duration in log domain (B, T)
-        :param torch.Tensor targets: groundtruth duration in linear domain (B, T)
-        """
-        # NOTE: outputs is in log domain while targets in linear
-        targets = torch.log(targets.float() + self.offset)
-        loss = self.criterion(outputs, targets)
-
-        return loss
-
-
-class DurationCalculator(torch.nn.Module):
-    """Duration calculator using teacher model
-
-    :param torch.nn.Module teacher_model: teacher auto-regressive Transformer
-    """
-
-    def __init__(self, teacher_model):
-        super(DurationCalculator, self).__init__()
-        if not isinstance(teacher_model, Transformer):
-            raise ValueError("teacher model should be the instance of e2e_tts_transformer.Transformer")
-        self.teacher_model = teacher_model
-        self.register_buffer("diag_head_idx", torch.tensor(-1))
-
-    def forward(self, xs, ilens, ys, olens):
-        """Calculate duration of each inputs
-
-        :param torch.Tensor xs: batch of padded character ids (B, Tmax)
-        :param torch.Tensor ilens: list of lengths of each input batch (B)
-        :param torch.Tensor ys: batch of padded target features (B, Lmax, odim)
-        :param torch.Tensor ilens: list of lengths of each output batch (B)
-        :return torch.Tensor: batch of durations (B, Tmax)
-        """
-        att_ws = self._calculate_attentions(xs, ilens, ys, olens)
-        # TODO(kan-bayashi): fix this issue
-        # this does not work in multi-gpu case. registered buffer is not saved.
-        if int(self.diag_head_idx) == -1:
-            self._init_diagonal_head(att_ws)
-        att_ws = att_ws[:, self.diag_head_idx]
-        durations = [self._calculate_duration(att_w, ilen, olen) for att_w, ilen, olen in zip(att_ws, ilens, olens)]
-
-        return pad_list(durations, 0)
-
-    @staticmethod
-    def _calculate_duration(att_w, ilen, olen):
-        return torch.stack([att_w[:olen, :ilen].argmax(-1).eq(i).sum() for i in range(ilen)])
-
-    def _init_diagonal_head(self, att_ws):
-        diagonal_scores = att_ws.max(dim=-1)[0].mean(dim=-1).mean(dim=0)  # (H * L,)
-        self.register_buffer("diag_head_idx", diagonal_scores.argmax())
-
-    def _calculate_attentions(self, xs, ilens, ys, olens):
-        att_dict = self.teacher_model.calculate_all_attentions(
-            xs, ilens, ys, olens, skip_output=True, keep_tensor=True)
-        return torch.cat([att_dict[k] for k in att_dict.keys() if "src_attn" in k], dim=1)  # (B, H*L, Lmax, Tmax)
-
-
-class LengthRegularizer(torch.nn.Module):
-    """Length regularizer module
-
-    :param float pad_value: value used for padding
-    """
-
-    def __init__(self, pad_value=0.0):
-        super(LengthRegularizer, self).__init__()
-        self.pad_value = pad_value
-
-    def forward(self, xs, ds, ilens, alpha=1.0):
-        """Apply length regularizer
-
-        :param torch.Tensor xs: input tensor with the shape (B, Tmax, D)
-        :param torch.Tensor ds: duration of each components of each sequence (B, T)
-        :param torch.Tensor ilens: batch of input lengths (B,)
-        :param float alpha: alpha value to control speed of speech
-        :return torch.Tensor: length regularized input tensor (B, T*, D)
-        """
-        assert alpha > 0
-        if alpha != 1.0:
-            ds = torch.round(ds.float() * alpha).long()
-        xs = [x[:ilen] for x, ilen in zip(xs, ilens)]
-        ds = [d[:ilen] for d, ilen in zip(ds, ilens)]
-        xs = [self._repeat_one_sequence(x, d) for x, d in zip(xs, ds)]
-
-        return pad_list(xs, self.pad_value)
-
-    def _repeat_one_sequence(self, x, d):
-        """Repeat each frame according to duration
-
-        >>> x = torch.tensor([[1], [2], [3]])
-        tensor([[1],
-                [2],
-                [3]])
-        >>> d = torch.tensor([1, 2, 3])
-        tensor([1, 2, 3])
-        >>> self._repeat_one_sequence(x, d)
-        tensor([[1],
-                [2],
-                [2],
-                [3],
-                [3],
-                [3]])
-
-        :param torch.Tensor x: input tensor with the shape (T, D)
-        :param torch.Tensor d: duration of each frame of input tensor (T,)
-        :return torch.Tensor: length regularized input tensor (T*, D)
-        """
-        return torch.cat([x_.repeat(int(d_), 1) for x_, d_ in zip(x, d) if d_ != 0], dim=0)
-
-
-class DurationPredictor(torch.nn.Module):
-    """Duration predictor module
-
-    :param int idim: input dimension
-    :param int n_layers: number of convolutional layers
-    :param int n_chans: number of channels of convolutional layers
-    :param int kernel_size: kernel size of convolutional layers
-    :param float dropout_rate: dropout rate
-    :param float offset: offset value to avoid nan in log domain
-    """
-
-    def __init__(self, idim, n_layers=2, n_chans=384, kernel_size=3, dropout_rate=0.1, offset=1.0):
-        super(DurationPredictor, self).__init__()
-        self.offset = offset
-        self.conv = torch.nn.ModuleList()
-        for idx in range(n_layers):
-            in_chans = idim if idx == 0 else n_chans
-            self.conv += [torch.nn.Sequential(
-                torch.nn.Conv1d(in_chans, n_chans, kernel_size, stride=1, padding=(kernel_size - 1) // 2),
-                torch.nn.ReLU(),
-                LayerNorm(n_chans, dim=1),
-                torch.nn.Dropout(dropout_rate)
-            )]
-        self.linear = torch.nn.Linear(n_chans, 1)
-
-    def forward(self, xs, x_masks=None):
-        """Calculate duration predictor forward propagation
-
-        :param torch.Tensor xs: input tensor (B, Tmax, idim)
-        :param torch.Tensor x_masks: mask for removing padded part (B, Tmax)
-        :return torch.Tensor: predicted duration tensor in log domain (B, Tmax)
-        """
-        xs = xs.transpose(1, -1)  # (B, idim, Tmax)
-        for idx in range(len(self.conv)):
-            xs = self.conv[idx](xs)  # (B, C, Tmax)
-        xs = self.linear(xs.transpose(1, -1)).squeeze(-1)  # (B, Tmax)
-
-        if x_masks is not None:
-            xs = xs.masked_fill(x_masks, 0.0)
-
-        return xs
-
-    def inference(self, xs, x_masks=None):
-        """Inference duration
-
-        :param torch.Tensor xs: input tensor with tha shape (B, Tmax, idim)
-        :param torch.Tensor x_masks: mask for removing padded part (B, Tmax)
-        :return torch.Tensor: predicted duration in linear domain with the shape (B, Tmax)
-        """
-        xs = xs.transpose(1, -1)  # (B, idim, Tmax)
-        for idx in range(len(self.conv)):
-            xs = self.conv[idx](xs)  # (B, C, Tmax)
-        xs = self.linear(xs.transpose(1, -1))  # (B, Tmax, 1)
-        xs = torch.clamp(torch.round(torch.exp(xs) - self.offset), min=0)  # avoid negative value
-        xs = xs.squeeze(-1).long()
-
-        if x_masks is not None:
-            xs = xs.masked_fill(x_masks, 0)
-
-        return xs
